+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.12.0 (2025-02-21)
 
 - feat: automatic markdown linting via `markdownlint`.
 - feat: lazy extension dependencies.